refactor(notes): extract keyword management into useNotesKeywords hook
Move 24 useState/useRef calls, 8 useMemo, 20 useCallback, and 1 useEffect
for keyword picker/manager/suggestions into a dedicated custom hook.
NotesManagerPage: 6055 → 5559 lines. All 137 tests pass.
